renamed video.ts to video-tracker.ts
